requirements.txt
streamlit
langchain
langchain-community
chromadb
pypdf
python-dotenv
requests